backend review update

+++ backend/src/main/java/com/example/backend/repo/DoctorRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository //Data Access Object layer which interacts with the table.
-public interface DoctorRepo extends JpaRepository<doctors,Integer>{
+public interface DoctorRepo extends JpaRepository<doctors,Long>{
     //spring data jpa provides the readymade CRUD operations
 }
